GskGLDriver: Use float for texture size in create_texture
Render nodes can end up with bounds < 1 since they are floats, and the
implicit cast to int ends up creating a texture with 0 width or height.
Use ceil() instead in create_texture so we don't have to do that on the
caller side everywhere.

+++ gsk/gskgldriverprivate.h
@@ -28,8 +28,8 @@ int             gsk_gl_driver_get_texture_for_texture   (GskGLDriver     *driver
                                                          int              min_filter,
                                                          int              mag_filter);
 int             gsk_gl_driver_create_texture            (GskGLDriver     *driver,
-                                                         int              width,
-                                                         int              height);
+                                                         float            width,
+                                                         float            height);
 int             gsk_gl_driver_create_vao_for_quad       (GskGLDriver     *driver,
                                                          int              position_id,
                                                          int              uv_id,
